using unsigned only

+++ include/common.h
@@ -48,7 +48,7 @@
 	/**@}*/
 	
 	/* Forward definitions. */
-	extern double *create_tasks(unsigned, unsigned);
+	extern unsigned *create_tasks(unsigned, unsigned);
 	
 	/* Forward definitions. */
 	extern const char *distributions[NDISTRIBUTIONS];
